+++ src/entropy.rs
@@ -1,3 +1,4 @@
+use generic_array::{ArrayLength, GenericArray};
 use rand_core::{CryptoRng, RngCore, SeedableRng};
 use rand_hc::Hc128Rng;
 
@@ -20,8 +21,8 @@ impl CsRng {
     /// Generate a vector of random bytes with the given length.
     ///
     /// - `len` : number of random bytes to generate
-    pub fn generate_random_bytes(&mut self, len: usize) -> Vec<u8> {
-        let mut bytes = vec![0_u8; len];
+    pub fn generate_random_bytes<N: ArrayLength<u8>>(&mut self) -> GenericArray<u8, N> {
+        let mut bytes = GenericArray::<u8, N>::default();
         self.rng.fill_bytes(&mut bytes);
         bytes
     }
@@ -56,16 +57,17 @@ impl CryptoRng for CsRng {}
 #[cfg(test)]
 mod test {
 
-    use super::*;
+    use crate::entropy::CsRng;
+    use generic_array::typenum::{Unsigned, U1024};
 
     #[test]
     fn test_random_bytes() {
         let mut cs_rng = CsRng::default();
-        let size: usize = 1024;
-        let random_bytes_1 = cs_rng.generate_random_bytes(size);
-        assert_eq!(size, random_bytes_1.len());
-        let random_bytes_2 = cs_rng.generate_random_bytes(size);
-        assert_eq!(size, random_bytes_1.len());
+        type N = U1024;
+        let random_bytes_1 = cs_rng.generate_random_bytes::<N>();
+        assert_eq!(<N as Unsigned>::to_usize(), random_bytes_1.len());
+        let random_bytes_2 = cs_rng.generate_random_bytes();
+        assert_eq!(<N as Unsigned>::to_usize(), random_bytes_1.len());
         assert_ne!(random_bytes_1, random_bytes_2);
     }
 }