+++ src/lib.rs
@@ -18,17 +18,24 @@ pub mod entropy;
 pub mod kdf;
 pub mod symmetric_crypto;
 
-use zeroize::Zeroize;
+use zeroize::{Zeroize, ZeroizeOnDrop};
 
 pub use crate::error::CryptoCoreError;
 
+pub mod reexport {
+    pub use generic_array;
+    // reexport `rand_core` so that the PRNG implement the correct version of the traits
+    pub use rand_core;
+}
+
 /// Trait defining a cryptographic key.
-pub trait KeyTrait: PartialEq + Eq + Send + Sync + Sized + Clone + Zeroize {
+pub trait KeyTrait: PartialEq + Eq + Send + Sync + Sized + Clone + Zeroize + ZeroizeOnDrop {
     /// Number of bytes in the serialized key.
-    const LENGTH: usize;
+    type Length: Eq + generic_array::ArrayLength<u8>;
 
     /// Convert the given key into a vector of bytes.
-    fn to_bytes(&self) -> Vec<u8>;
+    #[must_use]
+    fn to_bytes(&self) -> generic_array::GenericArray<u8, Self::Length>;
 
     /// Convert the given bytes into a key. An error is returned in case the
     /// conversion fails.