+++ src/symmetric_crypto/metadata.rs
@@ -59,6 +59,11 @@ impl<'a> BytesScanner<'a> {
             Some(remainder)
         }
     }
+
+    /// Return `true` if there still are some bytes to read.
+    pub fn has_more(&self) -> bool {
+        self.start < self.bytes.len()
+    }
 }
 
 /// Metadata encrypted as part of the header