+++ src/symmetric_crypto/mod.rs
@@ -10,9 +10,11 @@ mod block;
 mod metadata;
 
 pub use block::Block;
+pub use metadata::BytesScanner;
 pub use metadata::Metadata;
 
 use crate::{CryptoCoreError, KeyTrait};
+use generic_array::GenericArray;
 use nonce::NonceTrait;
 use rand_core::{CryptoRng, RngCore};
 use std::vec::Vec;
@@ -21,6 +23,10 @@ use std::vec::Vec;
 pub trait SymKey: KeyTrait {
     /// Convert the given key into a byte slice.
     fn as_bytes(&self) -> &[u8];
+
+    /// Convert the given bytes into a key.
+    #[must_use]
+    fn from_bytes(bytes: GenericArray<u8, Self::Length>) -> Self;
 }
 
 /// Defines a symmetric encryption scheme. If this scheme is authenticated,